aggiunti file visualizzazione heatmap
